+++ internal/service/platform/usergroup/daa_source_usergroup_test.go
@@ -103,6 +103,30 @@ func testAccDataSourceUserGroupByName(id string, name string) string {
 			name = "%[2]s"
 			org_id = harness_platform_project.test.org_id
 			project_id = harness_platform_project.test.id
+			linked_sso_id = "linked_sso_id"
+			externally_managed = false
+			users = ["pTaX1uGFQiSUm5_5lAM8kw"]
+			notification_configs {
+				type = "SLACK"
+				slack_webhook_url = "https://google.com"
+			}
+			notification_configs {
+				type = "EMAIL"
+				group_email = "email@email.com"
+			}
+			notification_configs {
+				type = "MSTEAMS"
+				microsoft_teams_webhook_url = "https://google.com"
+			}
+			notification_configs {
+				type = "PAGERDUTY"
+				pager_duty_key = "pagerDutyKey"
+			}
+			linked_sso_display_name = "linked_sso_display_name"
+			sso_group_id = "sso_group_id"
+			sso_group_name = "sso_group_name"
+			linked_sso_type = "SAML"
+			sso_linked = true
 		}
 
 		data "harness_platform_usergroup" "test" {