+++ internal/service/platform/usergroup/data_source_usergroup.go
@@ -19,7 +19,85 @@ func DataSourceUserGroup() *schema.Resource {
 
 		ReadContext: dataSourceUserGroupRead,
 
-		Schema: map[string]*schema.Schema{},
+		Schema: map[string]*schema.Schema{
+			"linked_sso_id": {
+				Description: "The SSO account ID that the user group is linked to.",
+				Type:        schema.TypeString,
+				Computed:    true,
+			},
+			"externally_managed": {
+				Description: "Whether the user group is externally managed.",
+				Type:        schema.TypeBool,
+				Computed:    true,
+			},
+			"users": {
+				Description: "List of users in the UserGroup.",
+				Type:        schema.TypeSet,
+				Computed:    true,
+				Elem: &schema.Schema{
+					Type: schema.TypeString,
+				},
+			},
+			"notification_configs": {
+				Description: "List of notification settings.",
+				Type:        schema.TypeSet,
+				Optional:    true,
+				Elem: &schema.Resource{
+					Schema: map[string]*schema.Schema{
+						"type": {
+							Description: "Can be one of EMAIL, SLACK, PAGERDUTY, MSTEAMS",
+							Type:        schema.TypeString,
+							Computed:    true,
+						},
+						"slack_webhook_url": {
+							Description: "Url of slack webhook",
+							Type:        schema.TypeString,
+							Computed:    true,
+						},
+						"microsoft_teams_webhook_url": {
+							Description: "Url of Microsoft teams webhook",
+							Type:        schema.TypeString,
+							Computed:    true,
+						},
+						"pager_duty_key": {
+							Description: "Pager duty key",
+							Type:        schema.TypeString,
+							Computed:    true,
+						},
+						"group_email": {
+							Description: "Group email",
+							Type:        schema.TypeString,
+							Computed:    true,
+						},
+					},
+				},
+			},
+			"linked_sso_display_name": {
+				Description: "Name of the linked SSO.",
+				Type:        schema.TypeString,
+				Computed:    true,
+			},
+			"sso_group_id": {
+				Description: "Identifier of the userGroup in SSO.",
+				Type:        schema.TypeString,
+				Computed:    true,
+			},
+			"sso_group_name": {
+				Description: "Name of the SSO userGroup.",
+				Type:        schema.TypeString,
+				Computed:    true,
+			},
+			"linked_sso_type": {
+				Description: "Type of linked SSO",
+				Type:        schema.TypeString,
+				Computed:    true,
+			},
+			"sso_linked": {
+				Description: "Whether sso is linked or not",
+				Type:        schema.TypeBool,
+				Computed:    true,
+			},
+		},
 	}
 
 	helpers.SetMultiLevelDatasourceSchema(resource.Schema)